refactor(TextLayer): move `TextLayer` into its own file and call `setLayerHeight()` on element updates
